feat(remote-sync): refactor error handling and update imports for consistency
